* 'develop' of https://gitlab.com/jusfoun-bigdata/hookah:
  数据插入mongo
  字段调整
  字段调整
  商品分类完善
  no message
  修改注resource部分
  字段 属性更改
  农行K宝接口调用
  后台数据演示
  之前包建错了 ^_^
  dx
  dx
  根据userId查询payAccountRecordId

# Conflicts:
#	hookah-console-server/pom.xml
#	hookah-oauth2server/pom.xml
#	hookah-website/pom.xml

+++ hookah-core/src/main/java/com/jusfoun/hookah/core/domain/PayBankCard.java
@@ -3,6 +3,7 @@ package com.jusfoun.hookah.core.domain;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.jusfoun.hookah.core.generic.GenericModel;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.GeneratedValue;
@@ -15,7 +16,7 @@ public class PayBankCard extends GenericModel {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    private Integer payAccountId;
+    private Long payAccountId;
 
     private String userId;
 
@@ -49,11 +50,11 @@ public class PayBankCard extends GenericModel {
         this.id = id;
     }
 
-    public Integer getPayAccountId() {
+    public Long getPayAccountId() {
         return payAccountId;
     }
 
-    public void setPayAccountId(Integer payAccountId) {
+    public void setPayAccountId(Long payAccountId) {
         this.payAccountId = payAccountId;
     }
